src/data_agent/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-dataform==0.8.0
vertexai==0.4.1
langchain-core==0.1.23
langchain==0.1.7
langchain-google-vertexai==0.0.5
langgraph==0.0.24
pydantic==1.10.13
PyYAML==6.0.1